frontend/requirements.txt
pinecone
neo4j
boto3
streamlit
extra-streamlit-components
requests
python-dotenv